OI Convert CSV: Fix import of fractional quantities.
git-svn-id: https://openintents.googlecode.com/svn/trunk@3138 72b678ce-9140-0410-bee8-679b907dd61a

+++ ShoppingList/src/org/openintents/provider/Shopping.java
@@ -1087,7 +1087,7 @@ public abstract class Shopping {
 	 */
 	public static long addItemToList(Context context, final long itemId,
 			final long listId,	final long status, final long priority, 
-			final long quantity) {
+			final double quantity) {
 		long id = -1;
 		Cursor existingItems = context.getContentResolver()
 				.query(Contains.CONTENT_URI, new String[] { Contains._ID },
